requirements.txt
Flask
gunicorn
pymongo
dnspython
pyjwt
werkzeug
flask-cors
prometheus_client
python-dotenv
scikit-learn
numpy
pandas
